requirements.txt
flask==3.0.2
mysql-connector-python==8.3.0
psycopg2-binary==2.9.9
pika==1.3.2
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pandas==2.2.1